pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-mc"
version = "0.1.0"
description = "Monte Carlo risk & scenario simulation toolkit for crypto time series"
authors = [{name="You", email="you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.25",
    "scipy>=1.11",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "statsmodels>=0.14",
    "arch>=6.4",
    "matplotlib>=3.8"
]

[project.scripts]
crypto-mc = "crypto_mc.cli:main"
